requirements.txt
pandas
pyyaml
scikit-learn
transformers>=4.30.0
datasets>=2.14.0
torch>=2.0.0
pyyaml>=6.0
sacrebleu>=2.3.1
rouge-score>=0.1.2
dvc>=3.0.0
datasets